+++ djrill/__init__.py
@@ -1,2 +1,3 @@
 from ._version import __version__, VERSION
-from .exceptions import MandrillAPIError, MandrillRecipientsRefused, NotSupportedByMandrillError
+from .exceptions import (MandrillAPIError, MandrillRecipientsRefused,
+                         NotSerializableForMandrillError, NotSupportedByMandrillError)